Добавлены правки по комментариям
