+++ src/test/resources/features/parameterizedTest.feature
@@ -2,7 +2,7 @@
 #encoding:UTF-8
 
   @param
-  Функционал: Параметризированный запуск тестов
+  Функционал: Проверка цены товара
 
     Структура сценария: Проверка цены книг <gadgetName>
       Допустим открыта страница "https://demowebshop.tricentis.com/books"
