+++ src/test/resources/features/searchItemsTest.feature
@@ -3,11 +3,9 @@
 
 @search
 Функционал: Поиск товара
-  Какм пользователь
-  Я хочу искать товары
 
   Предыстория:
-    Допустим открыта страница "https://demowebshop.tricentis.com/"
+    Допустим открыта домашняя страница "https://demowebshop.tricentis.com/"
 
   Сценарий: Поиск ноутбуков
     Когда в поле поиска введено значение "laptop"
